refactoring redux
