+++ package.json
@@ -11,6 +11,7 @@
     "react-redux": "^7.2.2",
     "react-scripts": "4.0.2",
     "redux": "^4.0.5",
+    "redux-thunk": "^2.4.1",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
