+++ src/store/types.js
@@ -1,0 +1,7 @@
+export const LOAD_REVIEWER_SUCCESS = 'LOAD_REVIEWER_SUCCESS';
+export const LOAD_REVIEWER_FAILURE = 'LOAD_REVIEWER_FAILURE';
+
+export const SET_SETTINGS = 'SET_SETTINGS';
+
+export const LOAD_USER_SUCCESS = 'LOAD_USER_SUCCESS';
+export const LOAD_USER_FAILURE = 'LOAD_USER_FAILURE';
